# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Support Requirement System ",
        "version": "1.0.0",
        "description": "A simple web application where [Frontend](https://github.com/adityajain-Infobeans/Support-Requirement-System-Frontend) is created using Vue and [Backend](https://github.com/adityajain-Infobeans/Support-Requirement-System-Backend) is created on node using express. Frontend will communicate with Backend via REST APIs for authentication and database operations",

        "contact": {
            "name": "Aditya Jain",
            "url": "https://aadityajain.dev",
            "email": "adityan.jain@infobeans.com"
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Development server"
        },
        {
            "url": "https://infobeans-support.herokuapp.com/",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/requirement/{requirement_id}": {
            "get": {
                "summary": "Get requirement data",
                "tags": ["Requirement"],
                "description": "Will return requirement data associated with provided id.",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "requirement_id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "description": "Requirement id (required)"
                    }
                ],

                "responses": {
                    "200": {
                        "description": "Requirement data successfully retrieved"
                    },
                    "406": {
                        "description": "Invalid requirement id"
                    },
                    "500": {
                        "description": "Error while querying requirement"
                    }
                }
            },
            "delete": {
                "summary": "Delete requirement entry from database",
                "tags": ["Requirement"],
                "description": " Delete the requirement with provided requirement id from database",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "requirement_id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "description": "Requirement id (required)"
                    }
                ],

                "responses": {
                    "200": {
                        "description": "Requirement deleted successfully"
                    },
                    "400": {
                        "description": "Requirement id is not provided"
                    },
                    "404": {
                        "description": "Id doesn't exist in database"
                    },
                    "503": {
                        "description": "Error while querying data"
                    }
                }
            },
            "put": {
                "summary": "Update requirement entry in database.",
                "tags": ["Requirement"],
                "description": "Update the requirement entry with provided id in database",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "requirement_id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "description": "Requirement id (required)"
                    }
                ],
                "requestBody": {
                    "description": "Create a new entry with provided data in database",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "status",
                                    "priority",
                                    "subject",
                                    "description",
                                    "client_id"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string"
                                    },
                                    "priority": {
                                        "type": "string"
                                    },
                                    "contact": {
                                        "type": "integer"
                                    },
                                    "subject": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "client_id": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },

                "responses": {
                    "200": {
                        "description": "Requirement successfully updated"
                    },
                    "400": {
                        "description": "Requirement id is not provided"
                    },
                    "404": {
                        "description": "Id doesn't exist in database"
                    },
                    "500": {
                        "description": "Error while updating requirement"
                    }
                }
            }
        },
        "/requirement/": {
            "get": {
                "summary": "Get all requirement data",
                "tags": ["Requirement"],
                "description": "It will return all requirements data",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],

                "responses": {
                    "200": {
                        "description": "Employee Found"
                    },
                    "404": {
                        "description": "No requirement found in database"
                    },
                    "503": {
                        "description": "Error while querying requirements"
                    }
                }
            },
            "post": {
                "summary": "Create new requirement entry in database.",
                "tags": ["Requirement"],
                "description": "Create a new entry with provided data in database",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "description": "Create a new entry with provided data in database",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "status",
                                    "priority",
                                    "subject",
                                    "description",
                                    "client_id"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string"
                                    },
                                    "priority": {
                                        "type": "string"
                                    },
                                    "contact": {
                                        "type": "integer"
                                    },
                                    "subject": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "client_id": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },

                "responses": {
                    "200": {
                        "description": "Requirement created successfully"
                    },
                    "400": {
                        "description": "Parameter missing"
                    },
                    "500": {
                        "description": "Error while querying requirement"
                    }
                }
            }
        },
...
    }
}